+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -547,6 +547,14 @@ export class NgxLocationPickerComponent
           coords = this.locationPickerHelper.convertLambertToWgs84Coordinates(coords);
           searchValue = `${coords.x}, ${coords.y}`;
         }
+        this.addMapMarker([coords.x, coords.y], null, false, false,  this.createMarker(
+          "var(--THEME1-800)",
+          "ai-pin-3",
+          "var(--SPACER)",{
+          top: "-0.75rem",
+          left: "-0.25rem",
+        }
+      ));
         this.writeValue(tempLocation);
       }
 
@@ -938,6 +946,7 @@ export class NgxLocationPickerComponent
    * @param location the selected location
    * @param keepGeometry whether or not to remove existing geometry
    * @param keepMarker whether or not to remove existing marker
+   * @param marker the marker styling itself
    */
   private addMapMarker(
     coords,