+++ CHANGELOG.md
@@ -17,8 +17,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Updated location viewer dependency
-
+- Updated location viewer dependency.
+- Added 'actual location' marker for the 'What is here' feature.
 
 ## [7.2.0]